Convert to autotools support
This doesn't convert the 96BoardsConfig code, but that will be going
away in an upcoming change that will base the library on libsoc.

The examples directory was changed to automake, but is not automatically
recursed into this. This allows people to optionally work with it in a
similar way they have in the past:
 cd ./example
 make
 sudo ./blink

Signed-off-by: Andy Doan <andy.doan@linaro.org>